+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changes
 
+- Added `passiveXpModifier` gamerule to modify the passive XP drop rate when `alwaysDropXp` is enabled.
+
 ## 5.0.0
 
 ### Changes